requirements/requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"

apache-airflow-providers-snowflake==4.0.2
apache-airflow-providers-mysql==4.0.0
apache-airflow-providers-sftp==4.2.1
apache-airflow-providers-zendesk==4.2.0
apache-airflow-providers-smtp==1.1.0
pandas==1.5.2
openpyxl==3.1.2
boto3==1.26.51
requests==2.28.2re

